account: Add the Account API client code
The Accounts API is a large API, because accounts include a _lot_ of
information.

To start, `AccountClient` represents the Account API.  It takes a
MAISClient, and stores things like caches.  It has a method to fetch
accounts, and—thanks to `AccountView`—a way to filter account lookups
(for example, to only return active accounts).

Information about individual accounts is stored in `Account`, along with
the code that actually fetches account information from the API.  Core
to this is `AccountService`, which stores service-specific information
that is associated with the account.

This is one big commit of three files, because the main module files
imports the others.

In addition to the `AccountView` filters included here, later on I'll
commit code to validate a set of account IDs in one operation.