Define a default timeout for Requests
You might assume that Requests Sessions allow you to set a timeout in a
session, but that's not the case!  It has been asked for many times, as
old as https://github.com/psf/requests/issues/1130 and as recently as
https://github.com/psf/requests/issues/3054.

Since timeouts have to be set in the reuqest, I'm defining a new
property of `MAISClient`, to hold the default timeouts.

So far, the only Requests call is made in the `Account` class, so
there's only one place that needed a timeout added.

This was caught in MyPy testing!