Include a support script to make test version numbers
PyPI doesn't normally accept pushing the same package version twice.
This isn't really a problem in production PyPI (since pushing the same
version number twice is _extremely_ unusual), this is a problem on test
PyPI, since we package a new test release on every push.

To allow these test pushes, this script modifies the version number
before the distributions are built.  Version numbers are transformed
like so:

    1.2.0 --> 1.1.99.datetime
    - or -
    1.2.1 --> 1.2.0.99.datetime

`datetime` is replaced with the `YYYYMMDDHHMMSS` UTC datetime.  That
does mean there is a race condition if two pushes are happening around
the same time, but that should be a rare event.

To be clear, this transformation _only_ happens when pushing to Test
PyPI.  The `VERSION` file should be bumped to the next version as soon
as a version is released.