Add tests to the GitHub workflow
Tests are added as a new step in the workflow.

Although Python 3.9 is the only supported version right now, a
python-version matrix is being used for tests, so that future Python
versions can be tested easily.  The linting and publish jobs will
continue to run on the latest Python version we happen to support.

Both publish jobs are updated to only run when tests pass.