AccountService: Fix typing for _get_settings
Tuples are handled differently from other collection types, when it
comes to typing.  Special care is needed to indiate "An empty tuple", or
"A tuple of variable # of items, where each is the same type".

One beneficial side-effect of this is that we no londer need to pass
tuple parameters that are empty.  They'll be automatically set to the
empty tuple at the time of the call.