account: Type cleanups around _get_settings
The type-checking around _get_settings and _get_setting needed a fair
bit of cleanup.  The parsed JSON we get can contain different types of
values, depending on the key.

_get_setting's type was updated, as we only get the settings list, so we
don't need to deal with the full settings type.

The return type of _get_settings was changed to use the Set as a return
type for multi-valued settings, instead of Tuple (we were using a set
anyway).

At the start of _get_settings, MyPy was getting confused about the
allowed types for keys of `result`.  Explicitly setting the type during
declaration fixed that.

Finally, I needed to add at least one isinstance check, to confirm type
before doing a set construction.