AccountClient: custom_session is for custom sessions
When a client wants to set a custom session, use `custom_session`
instead of `session`, to stop MyPy complaining.

In AccountClient, `session` is going to be a `requests.Session`, but
only once initialization is complete.  It is meant to be optional to the
user; they can provide a `Session` instance for us to use, but it's not
required.  Unfortunately, if we mark it as optional, MyPy complains when
we access the session without using it.

Using `typing.cast(...)` (in `__post_init__`) doesn't work.  The casting seems
to remain within the method.