account: Define _from_json, and set type on known_services
All of the subclasses of AccountService have a _from_json classmethod.
But the base AccountService class doesn't, so we needed to define it in
order for typechecking to work.

Note that this should really be marked as an abstract method in the base
class, but doing so triggers https://github.com/python/mypy/issues/5374;
until that's fixed, and we switch to that version, we need to manually
raise an exception.

Also, the `known_services` dict needed to have an explicit type
annotation set, since the value type could not be automatically
inferred: We need the value type to be the superclass type.