MyPy: Switch to package-discovery, instead of file-discovery
This switches to using MyPy's package-discovery mechanism, having it
enumerate and then type-check everything within and under
`stanford.mais`.

(`stanford.mais` is used instead of `stanford` in case anything else
starts using that namespace.)

This is better than listing a directory to check, because it ensures
that all modules have been discovered, and their relationships laid out.

Unfortunately, although it's possible to provide a list of
files/directories in config, you can't do this with modules/packages.
I've requested support for this in https://github.com/python/mypy/issues/10728