account: Add support for bulk validation of SUNetIDs.
This adds the `stanford.mais.account.validate` module, which contains
the `validate` function.  The function takes a list of SUNetIDs (for
people, not functional accounts), either as a string or a
tuple/set/list.  It returns an `AccountValidationResults` named tuple,
categorizing the SUNetIDs into full (or full-sponsored), base (or
base-sponsored), inactive, and unknown.  It's a big convenience for
users who just want to validate a list of SUNetIDs!

This uses the `only_people` Account View to make the code really simple.