Mark modules as fully-typed, per PEP 561.
All of the Python code in this package is fully typed.  That's great
(and essential), but typechecking tools like MyPy weren't able to say if
it was OK to use the typing info in these modules.

The way for indicating this is PIP 561
(https://www.python.org/dev/peps/pep-0561/), which for us means putting
a `py.typed` file at the root of each module we provide.  The file can
be empty; it's the presence of the file that provides the signal to
type-checkers.

This also modifies `MANIFEST.in` to ensure the `py.typed` files are
bundled into distributions.

Finally, `zip_safe` is disabled in `setup.cfg`.  This ensures that,
when someone tries to install this software using `setup.py install`, it
will not attempt to produce an egg file.  This is needed because MyPy
does not look inside egg files when searching for module type
information (see https://github.com/python/mypy/issues/8802 for more
details).  People shouldn't be using `setup.py` directly anymore anyway,
packages should be installed using `pip` (for example, `pip install
stanford-mais`, or `pip install .` if installing from a Git clone).

Closes: #1