Add config for Read The Docs
Read the Docs can build without a configuration file, but if you want to
do anything other than the defaults, you need to provide one.

In our case, we want to build the docs using Python 3.9.  So, we bring
in a configuration file.

This might fix #1, but that can't be confirmed until docs build.