README: Add Workgroup example, and update features
With the code for the Workgroup API being closer to done, the README can
now be updated with an example of how the SDK can be used.

Also, the list of what the SDK supports has changed:  Searching and
deleting is now supported.  Integrations (which was previously not
mentioned) is not supported.