Add an initial version of the Workgroup SDK
This adds an SDK for MaIS' Workgroup API!

This is probably the largest blob of code that will be added, since
(unlike other MaIS APIs) the Workgroups API allows reading and writing.
It includes the following:

* Creating & fetching workgroups, and refreshing already-fetched
  workgroups
* Searching for workgroups by name (with wildcard support)
* Searching for workgroups by person, certificate, or nested workgroup
* Deleting workgroups
* Accessing and changing workgroup properties, members, and
  administrators

WARNING: Searching for workgroups by member is broken right now, because
of a mismatch in the API's documentation and what is actually returned
when searching.

One thing that is not currently included is anything related to
workgroup integrations.