tests: Change Requests mocking to responses
This changes to using the Responses package for mocking of calls to the
Workgroups API.

As part of the change, I also change up how the PyTest fixtures are
created for the service clients: Instead of being session-scoped
fixtures, they are now normal fixtures.  This change means that clients
(including all mocking) will be set up and torn down before & after each
test.

I'm not sure why, but it seems like Responses mocks used in a PyTest
fixture only survive for a single test; a session-scoped fixture loses
the mocks after the first test and the second test.