Move Session management to MAISClient
Right now, each client (AccountClient, WorkgroupClient, etc.) is
responsible for managing its own Requests Session.  And if a user
wants to use a custom Requests Session (such as for testing), then they
need to provide it to each Client instance.  If they don't provide one,
then the "Session factory" in MAISClient is used instead.

In addition, there is inconsistent use of timeouts in code.  MAISClient
contains a default timeout, but it's up to callers to reference those
timeouts in their code.  Some call sites included the timeouts, but many
did not.

Really, Session creation and management should be taking place in one
location, in MAISClient.  That helps the individual client modules be
specific to only their code.

This commit implements the change by adding Session-creating code in
MAISClient.  The `session` method is removed, and code is added to
`__post_init__` to create the Session (if one is not already provided by
the user).  The documentation around `cert`, `key`, and
`default_timeout` is updated to reflect that they will be ignored if the
user provides a session.

Finally, in AccountClient and WorkgroupClient, all call sites have
references changed, pointing to the `MAISClient.session`.

TODO: Right now, timeouts are not being handled at all.  I probably need
to override one of the `Session` methods, but that'll be in a future
commit.