readthedocs: Update configuration
RTD is not building right now, as some of the configuration is old.  So,
this updates the configuration.

The build section is updated to use Ubuntu 24.04 LTS (the most recent
LTS that is currently available).  The Python version info moves into
the build section, with the remainder in the Python section.  And we no
longer use a test image for building.