Account Validation: Return frozensets
Within AccountValidationResults, use frozensets for all members except
`raw` and `raw_set`.  This is an additional step to ensure the results
cannot be accidentally modified by client code.  Typing should already
ensure this at a static level (as the members are typed as Collections
of strings); using frozensets enforces this at a dynamic level
(especially as AccountValidationResults is now a frozen dataclcass).

`raw` and `raw_set` are not changed, because AccountValidationResults
defines them as having the types of their inputs.

(Of course it's really difficult to make things truely immutable in
Python, but I think this goes pretty far!)