MAISClient: Add convenience method for UAT1
The new UAT1 environment is similar to UAT, except that it points to the
new Sequoa HR environment.  It exists for need to test against Sequoa;
you can change from calling `.uat()` to calling `.uat1()`.  Per MaIS,
the certificates you use with UAT should also work with UAT1.

NOTE: UAT1 does not provide APIs for Course or Student, so those are not
defined in `.uat1()`.  (Which doesn't matter anyway, since we don't have
any code for those services.)

BIGGER NOTE: UAT1 only provides a Workgroups API using the older 1.0
(XML-based) format.  The code in this package is designed for use with
the 2.0 (JSON-based) API.  That means that this package will not work
with UAT1 right now.