MAISClient: Close cert that was test-opened
While I was working on making warnings into errors in PyTest, I
encountered a ton of warnings, which all traced back to
`MAISClient.__post_init__()`: The file handle from
`self.cert.open()` was never caught, and so Python kept the cert file
open until the end of the program.  With the PyTest `-W error` setting,
this caused almost all tests to fail.

Now, the opening is wrapped in a context manager, which takes care of
immediately closing the opened file.