MAISClient: Rework timeout support
This is a big change to how timeouts are handled.
I mistakenly assumed that Requests Sessions supported a default timeout.
Turns out, that's not true: As made clear in
https://github.com/psf/requests/issues/1130,
https://github.com/psf/requests/issues/2856, and
https://github.com/psf/requests/issues/3054, Requests Sessions will not
be getting default-timeout support.

So, we now subclass the Session class, add `timeout` as an argument, and
override the `send` method to ensure a timeout (if set in the subclass)
is included.

As part of this, the `MAISClient` constructor code changes to
instantiate the subclass and set the timeout.  And our PROD/UAT/UAT1
class methods now take `timeout` as an argument.

In fact, the `MAISClient` has been changed, so `default_timeout` is now
`timeout`, and defaults to None.  It takes the same range of options as
Requests: Either `None`, a single float, or a two-float tuple.

Finally, a NamedTuple is defined, so folks can be clear about which
timeout is which.