Workgroup properties: Simplify from_str
Right now, the `from_str()` method in `WorkgroupVisibility` and
`WorkgroupFilter` do checks on the incoming string.  By changing the
Enum values from integers to strings, we can use the Enum constructor to
do our checking for us!

This change should not be visible to clients, since the Enum's value
should be opaque to them.