Clean up typing uses for Python 3.9
With the minimum-Python change from Python 3.8 to Python 3.9, a number
of changes can be made to how types are used.

In particular, types which had to be imported from `typing` (like
`typing.Set`) are now replaced with either the built-in type (like
`set`), or a type from `collections.abc` (like
`collections.abc.Mapping`).  For those types, the classes in `typing`
are deprecated, although Python isn't raising deprecation warnings for
them right now.

Also, the types on classmethods have been removed.

Finally, instead of `from typing import *`, only the needed items are
being imported.