Workflow: Run Bandit manually, for now
PyCQA's bandit-action does not support reading config from a TOML file
right now, so we cannot use it.  This changes the workflow to do
essentially what that Action does, but in a way that works for us.

Once https://github.com/PyCQA/bandit-action/issues/15 is implemented, we
can switch back to using that!