Workflows: Replace Actions tags with commit hashes
This change is in response to the tj-actions/changed-files supply-chain
attack: The attackers were able to change which commit an existing tag
pointed to, so all the GitHub Actions workflows pointing to the tag ran
the attacker's code.

The partial protection to this is to specify commit IDs, since commit
IDs are immutable.  It's not full protection, in case the Actions we
refer to are themselves referring to things by tag name.

As of this commit, the mapping of tag to commit hash is:

* actions/checkout@4.2.2 is 11bd71901bbe5b1630ceea73d27597364c9af683
* actions/setup-python@v5.6.0 is a26af69be951a213d495a4c3e4e4022e16d87065
* actions/upload-artifact@v4.6.2 is ea165f8d65b6e75b540449e92b4886f43607fa02
* github/codeql-action/*@v3.29.3 is d6bbdef45e766d081b84a2def353b0055f728d3e
* pypa/gh-action-pypi-publish@v1.12.4 is 76f52bc884231f62b9a034ebfe128415bbaabdfc