setup.cfg: Reduce requests version for install only
This reduces the version of requests that we need, from >=2.30.0 to
>=2.28.2.  But, when installing the package for testing (`pip install
.[test]`), still require >=2.30.0.

This brings us down to the version that comes installed in Debian
bookworm, which is now Debian oldstable.  I don't see any differences
between 2.28.2 and 2.30.0 that would affect us.  But, there is a small
risk introduced by being unable to test against 2.28.2.

See https://github.com/getsentry/responses/issues/635 for an explanation
of why responses needs requests 2.30.0 or later.

NOTE: requests versions older than 2.32.0 are vulnerable to
GHSA-9wx4-h78v-vm56 / CVE-2024-35195, but our use of requests does not
trigger the vulnerability: We do not use `verify=False` with requests.
For folks who have this as a problem, we support requests 2.32.0 and
later (anything below 3.0).