pyproject: Update how we depend on wheel
Only install wheel if we are running Python 3.11 or 3.12, and only
install 0.45.x.

In the past, setuptools needed the wheel package to implement the `bdist_wheel`
command.  But in setuptools 70.1.0, they "adopted" the implementation,
and the wheel command removed its implementation in wheel 0.46.0.

We support setuptools 66.1.1 for Debian bookworm (oldstable)
compatibility, but there is a risk that someone would have an old
setuptools (pre-70.1.0), but the latest wheel; that would break
the creation of wheels.

This commit implements a workaround to guard against that situation.

First, we assume that if you're running Python 3.13, you have a
setuptools that is newer than 70.1.0: Python 3.13 was released October
7, 2024; the setuptools available at that time was 75.1.0.

Next, if you are running Python 3.11 or 3.12, we require an older wheel
package, one that still implements bdist_wheel.  That way, you'll either
have a newer setuptools (that implements bdist_wheel directly), or
you'll have an older setuptoold and an old-enough wheel.