Move lots from setup.cfg to pyproject.toml
With us now requiring a Setuptools above 62.5.0, we can now take advantage of
`pyproject.toml`'s `[project]` section.  This commit moves everything we can
from `setup.cfg` to `pyproject.toml`.  The idea is, only things required
to be in `setup.cfg` should be there.

Once we can require Setuptools 68.1.0 or above, we can move stuff that goes
into the `[tool.setuptools]` section.  For now, those remain in `setup.cfg`.
Also, we keep `version` defined in `setup.cfg`, since we are having setuptools
pull the version number from a file.

We don't need `setup.cfg` for editable installs, so as soon as it's
empty, we can delete it.

Also, this commit removes the unused `options.entry_points` section, along with
some commented-out requirements.