Update test-push version logic
This changes the script that updates the `VERSION` file, when doing a
push to Test PyPi: It now appends a development-release version number.

The changes here are all based on the guidance from the Python Packaging User
Guide's section on Version speficiers.
https://packaging.python.org/en/latest/specifications/version-specifiers/

Folks who are doing Python version-number comparison should be treating
`a.b.c.devX` (for any value of `X`) as coming *before* `a.b.c` (and any
other forms of `a.b.c`, like `a.b.c-rc1`).  So, we no longer need to
'roll the version number back' to ensure it sorts properly.