Remove (almost) all Union types
Since we changed our Python requirement, we can now remove the reliance
on the Union type, replacing it with the now-standard `X | Y` syntax.

There is one exception: In `account.service`, MyPy is not properly
parsing the defintion of `OptionalTupleOfStrings` when the `X | Y` form
is being used.  So, a Union is left in place there.

See #8