MyPy: Enable explicit_package_bases
The explicit_package_bases setting tells MyPy that it should look in the
mypy_path for packages.  Doing this means we can tell MyPy to check the
repository root (`mypy .`), instead of needing to tell MyPy to check
things individually (`mypy tests/*.py && mypy -p stanford.mais`).

The only downside is that other Python files needed type-checking
information, as MyPy is now able to see them.  That is the reason for
the previous type-check commits!