account.service: Make AccountService abstract
This is another thing we can do now that we require MyPy 1.0.0+:
Previously, MyPy would raise an error if AccountService._from_json was
marked as an abstract method.  That's not a problem anymore, so
AccountService is now an Abstract Base Class, and
AccountService._from_json() is an abstract class method.