tests: Generate coverage data
This brings coverage data in to the test process.  Coverage records
which lines of code (everything in `src`) were executed by the PyTest
runs.  The results are recorded in a database, and output at the end of
the PyTest output.

`pyproject.toml` is updated to bring in pytest-cov (the minimum version
being what's in Debian oldstable), and Tox is updated to run coverage
with each pytest run.

Right now, the results are not being used, and the coverage database is
deleted at the end of each workflow run.  In the future, the coverage
results could be displayed somewhere!

Finally, the changelog is updated to reflect collection coverage.