account.AccountClient: Make only_* stacking work
Thanks to tests (that will be committed shortly!), I found a problem
with how I coded the `only_*()` views.  The view worked by returning an
AccountClient subclass, where `get()` is overridden to include a filter.

This works fine for a single filter, but if you try stacking, the only
`get()` methods that would run were the superclass's `get()`, followed
by the `get()` of the last filter in the chain.

Why is this?  It's because I was relying on subclassing, and not
recursion.  If I was using recursion, then a `get()` would propagate up
through all the AccountView instances, up to the topmost AccountClient,
with filters being applied as the returned Account goes back down.  But
subclasses do not work like that: There was just the AccountClient class
and the AccountView class, and so only one opportunity to apply filters.

The fix to this is to make filtering happen in AccountClient directly.
`_filters` is a new 'private' attribute, a frozenset of callables, the
same as AccountView's `_filter`.  Calling `only_*()` takes the current
instance's set of filters, adds a new one (making a new frozenset), and
returns a new AccountClient containing the expanded filters.