account tests: Use proper test data
This is a major rework to how test accounts are generated, and to some
of the tests themselves.

To start, instead of constructing JSON blobs, I took actual JSON blobs
from the Accounts API, and inserted our test data into them.  For the
values I did not already have, I made stuff up.  In other words:

* `owner` values were created using random data.

* `url` values were created using an `example.com` web site URL base.

This makes the mock file larger, but I think it's worth it, since we're
using raw data with real structure.

Since I was able to work off of real test data, that changed up some of
the tests.  In particular:

* For people, names are in 'last name first' form.

* If an inactive account had the kerberos service when it was active, it
  will still have the kerberos service when it is inactive, though the
  UID will be missing.

* If an inactive account had the leland service when it was active, it
  will still have the leland service when it is inactive.

* Shared mailboxes do not have the autoreply service.