account tests: Test accounts with unknown type & service
If an account has a type of service that we don't know about, then we
should continue working, but log a warning.

If an account is of a type that we've never seen before, raise an
exception.

We take the stronger action for unknown-type accounts because we cannot
be sure about the kind of data that the unknown-type account has.  Maybe
attributes that we think are required, are not required for this type of
account?  So, we take the safe route of raising an exception.