account.AccountService: Add not_inactive
The `is_active` service property is weird, when services could
potentially be frozen.  Sometimes you care if a service is frozen:
If someone's kerberos status is frozen, you want to block
logins.  Sometimes you do not care if a service is frozen: If a person's
kerberos status is frozen, you probably still want them to show up in a
directory listing.

So, for folks who want it, there is now `not_inactive`, which returns `True`
if the account is active or frozen.