pyproject coverage: Specify explicit packages & paths
Up to now, we were relying on coverage to autoatmically figure out what
we wanted to cover.  This was fine when running locally, but running in
Tox caused coverage to only report reults for the `tests` directory.

Coverage ignores everything in the standard library.  When testing
under Tox, Tox makes a source distribution of the code, and installs
that.  That causes coverage to ignore it.

When we test locally with an editable install, that installs a link to
our local `src` directory, which is enough for coverage to notice.

To fix this problem, we tell coverage to measure our packages,
regardless of where they're installed.  We then mention the tests
directory separately.

Finally, we tell coverage that `src/stanford/mais/account` is the the
same as any `stanford/mais/account` found elsewhere (such as via a
non-editable package install).  We do the same for `mais/client` and
`mais/workgroup`.

(Why the separate entries for account, client, and workgroup?  Because
there might be other stanford-namespace packages in the future!)