Workflow: Cache the Tox environment
This pulls the `.tox` directory from cache at the start of the run, and
pushes it at the end of the run.

The longest part of testing is the "Prepare Tox" step, which downloads &
installs the packages used for testing.  Right now, this does not use
any caching within GitHub Actions, so this step adds that!

By doing this, workflow numbers changed from…

|  Test Step  | Python 3.11 | Python 3.12 | Python 3.13 |
|-------------|-------------|-------------|-------------|
| Prepare pip |      3s     |      2s     |      3s     |
| Install pr… |      2s     |      2s     |      2s     |
| Prepare Tox |    2m45s    |    2m28s    |    2m25s    |

… to:

|  Test Step  | Python 3.11 | Python 3.12 | Python 3.13 |
|-------------|-------------|-------------|-------------|
| Cache tox … |      3s     |      3s     |      3s     |
| Prepare pip |      3s     |      2s     |      3s     |
| Install pr… |      2s     |      2s     |      2s     |
| Prepare Tox |    2m27s    |    2m11s    |     2m4s    |

* actions/cache@v4.2.4 is 0400d5f644dc74513175e3cd8d07132dd4860809

Thanks to https://github.com/lsst-sqre/safir/issues/2 for the example!