Workflow: Stop saving builds as artifacts
Previously, the action to publish pushes to PyPi (either PyPi or Test
PyPi) would first create an artifact, under the assumption that the
`pypa/gh-action-pypi-publish` Action needed this.  Turns out, that's not
true!

Not only that, but when tagging a release, both the PyPi *and* the Test
PyPi jobs would run; since they both use the same artifact name, that
would cause one of the two jobs to fail (you can't have two artifacts
with the same name in a single workflow run).