account.tests: Add mocks for account-changed API
In preparation for code which calls the "Get accounts changed in the
last X days" API, mock up some responses.

For each account JSON object, define an equivalent partial-account JSON
object.  And define a set of responses that return the partial-account
objects, when querying for inactive or active people or functional
accounts where the days-changed is 30.

Also, define an empty response for pending people where the days-changed
is 30.  Finally, define a couple of responses that intentionally return
errors.