workgroup: Have Workgroup use the cache
Workgroup's `.create()` and `.get()` methods now use the cache
maintained in the WorkgroupClient.  Newly-created workgroups are added
to the cache after creation.  When fetching a workgroup, we check if the
workgroup is in the cache, and only do the fetch if not.

Since the cache is a cache of weakrefs, we create & dereference them,
including checking that the deref worked.

Finally, when getting the workgroup, we lowercase the workgroup name at
the start of the method call.  This is to ensure our cache is
case-insensitive.