tests: Add test of WorkgroupClient
This adds tests for almost all of the code in stanford.mais.workgroup,
which includes WorkgroupClient and PartialWorkgroup.

Since some of the WorkgroupClient methods just pass the call to
Workgroup, this also ends up testing some Workgroup code.

The one thing not being tested right now is WorkgroupClient.create().
This also passes the call off to Workgroup, and fully testing it will
require a fair amount of code.  So, that will need to be done later.  In
the meantime, a skipped test is added.