workgroup.Workgroup: Update privgroup default & docs
There has been a change in the Workgroup API: Newly-created workgroups
have the privgroup property set to True.  So, the default value of the
Workgroup.create() method also changes.

Since we have not yet reached version 1.0 yet, we can change this
without a version-number bump.

In addition to changing the default, the docstring for the privgroup
property also changes.