workgroup.tests: Add JSON blobs for create
In preperation for testing the Workgroup.create() function, we need to
be able to test the creation of workgroups with properties set to
non-default values.  The first step is to create a load of JSON blobs,
one for a created workgroup with default properties, and one for each
different property value that we might see.

To reduce the number of JSON blobs a little, we only change one property
at a time.