workgroup.tests: Mock create calls
This adds mocks for different workgroup-create calls.

To start, there is a simple workgroup create call, with all properties
set to their defaults.

After that, a mock is added for each propery's not-default values.  That
is a *lot* of mocks, so to keep things from completely exploding, only
one property is changed in each mock operation.

Finally, we mock the response that is sent when attempting to create a
workgroup that either already exists, or that existed in the past (since
workgroups are never truely deleted).