tests: Add tests of Workgroup creation
This adds a file that tests everything related to Workgroup creation.
The first test checks the creation of a workgroup with default
properties.  Then, each attribute's non-default value is tested.

We also test that invalid names & descriptions generate exceptions, and
we check that upstream errors also generate exceptions.

Finally, since testing of workgroup creation is handled in a separate
file, we remove the skipped test from test_workgroup.py