account: Add PartialAccount
The PartialAccount class represents a "lite" account record.  Lite
accounts are returned when fetching a lite account (which we do not do),
and when searching for accounts that have changed status recently (which
we will soon do).

The class has the same structure as PartialWorkgroup: It's a frozen
dataclass with a method to fetch a full Account.

This commit also builds docs for PartialAccount.