workgroup.workgroup: Setting properties refreshes
When a workgroup's property is changed, the Workgroup API's response includes
the full workgroup record (with our changes applied).  Until now, we ignored
that, but that's not a good idea: For all we know, something major may have
happened to the workgroup (for example, it might have been deleted).  We should
take every opportunity to keep the user up-to-date.

So now, every time a workgroup property changes, we act as if the user
implicitly called `Workgroup.refresh()`, and update the entire Workgroup
instance.

This takes advantage of the change made in commit
c97ba8b57bb9d808839a4e25e671fcf72f4895a2.

As part of this change, the `Workgroup._update()` method is refactored:
Originally, it toook a dict of changes to make, because the API supports that.
But due to the way that we accept changes from users (as individual property
updates), we only ever make one change at a time.

So, the `Workgroup._update()` now accepts two parameters instead of one: It
accepts the name of the field to change, and the value.  This allows for a
little more type-checking!

Finally, the test cases for `Workgroup._update()` are changed to account for
change in parameters.