workgroup: A new exception for inactive workgroups
Workgroups are never completely "deleted".  When someone deletes a workgroup,
that actually sets it to "inactive".

When you try to fetch or modify an inactive workgroup, the Workgroup API
returns a 400 with a specific message.  We were not catching that, so to begin
doing so, create an appropriate exception.

This introduces the `WorkgroupDeleted` exception, a subclass of KeyError.  It
will be raised whenever we try to fetch or change an inactive workgroup.

The docs are also updated to include this new exception.