workgroup: Handle newly-inactive workgroups
As mentioned in commit 2afc6e65dea465033573b085c532c85b47fe893d, workgroups are
never "deleted", they simply go inactive.  This commit updates
stanford.mais.workgroup.workgroup and stanford.mais.workgroup.member — and
associated tests — to raise the new `WorkgroupDeleted` exception when a
workgroup is inactive.  The exception is raised when trying to fetch an
inactive workgroup, when trying to create a workgroup with the name of an
inactive workgroup, and when changing a workgroup that went inactive after we
fetched/created it.

Throughout workgroup.workgroup and workgroup.member, any code which calls out
to the Workgroup API has had its logic changed, in terms of how the 400
response code is handled.  Instead of immediately raising a ChildProcessError,
it looks for the special "Workgroup is inactive" message, and if found, raises
a WorkgroupDeletedError.

In workgroup.workgroup, the changes are more invasive.  If a workgroup has gone
inactive, before raising the WorkgroupDeletedError, we update the Workgroup
instance to look as if the client had called `Workgroup.delete()`.  This
involved making a new internal method `Workgroup._mark_deleted()` to to the
changes to the instance, replacing much of the code that `Workgroup.delete()`
used to run.

workgroup.member avoids an invasive code change because when it sees a
now-inactive workgroup, it hands off to `Workgroup._handle_refresh()`, which
does all the invasive work.

One test case needs to be modified as part of this change, and two test cases
need additional code, all because of the new exception.