account.tests: Mock a weird status code
Right now we do not handle the possibility that the server may return a status
code that we do not expect.  Before we can code for that, we need an unexpected
status code!  This mocks returning an unexpected status code for account fetch
and the "changed status in last X days" search.