workgroup.tests: Mock a weird status code
Right now we do not handle the possibility that the server may return a status
code that we do not expect.  Before we can code for that, we need an unexpected
status code!  This mocks returning an unexpected status code for searching by
name or cert, fetching a workgroup, creating a workgroup, setting a particular
description, getting a privgroup, and adding or removing a particular member.