Catch unexpected status codes
This changes all of the code—account and workgroup—which makes API calls.  The
code now has a catch for any undocumented response code (that is, HTTP response
codes that we should never see).  If we get an undocumented response code, we
raise a NotImplementedError to the caller.

Also of note, we now use `match` blocks for checking the status code, which
looks a lot cleaner than a series of `if` statements!  This was introduced in
Python 3.10; since we require Python 3.11+, we can use this.

Finally, we also update test cases for the new raisable exception, and update
function docs.