workgroup.tests: Mock property-changing responses
Add mocks covering changes to workgroup properties.  This also adds various
upstream response codes, triggered when changing the description to a special
value.

Each property change results in sending a JSON blob to the client.  In order
to minimze the amount of JSON blobs needed; a function is created that takes
test:1's blob, deserializes it, converts the specified property, and serializes
a new blob.