tests: Add tests for workgroup properties
This tests changing a workgroup's properties.  For boolean and enum properties,
we test each possible property value.  For the description, we test an OK
description.  We also test invalid properties, checking if the SDK raises
exceptions on them.

Finally, we use special description values to test handling of upstream
response codes.