workgroup: get_privgroup should not handle 404
This changes `Workgroup.get_privgroup()`, to make it not return a
specific exception when receiving a 404 from the API.

404 *is* a valid code that the API can return, which happens when the
caller fetches a privgroup for a workgroup that has never existed.  But,
clients can only get to the `get_privgroup()` method by fetching the
workgroup.  So, clients would encounter a 404 when first trying to fetch
the workgroup, not when trying to get the privgroup.

So, if `get_privgroup()` gets a 404, that is a very unusual situation.
We should not be treating this as something we should expect.  Instead,
we should be treating this as a "We got some reasponse code we never
expected to see" NotImplementedError.