workgroup: get_privgroup calls can_see_membership
This changes `Workgroup.get_privgroup()` to check `can_see_membership`
before making an API call.  If `can_see_membership` is False, the API
call will not be attempted and a PermissionError will be raised
immediately.

It is possible that we were not able to see a workgroup's contents when
we first fetched the workgroup, but now we can, either due to the
workgroup's visibility changing or our client cert being added as an
administrator.  In that case, the client will need to refresh the
workgroup before making this call.  However, that's probably going to be
rare.

NOTE: It is also possible that we *lost* access, which is something
we'll need to test.

Finally, this updates the method's documentation with notes about the
change.