workgroup: Add tests for get_privgroup
This adds tests for the `Workgroup.get_privgroup()` method, which was
the last method that was missing tests.  The new file covers tests for
fetching a privgroup, catching the short-circuiting PermissionError, and
catching upstream resonse codes.

Also, an existing test—checking for EOFError on a deleted instance—is
updated to also check `get_privgroup()`.