tests: Skip coverage on test_workgroup_member
This skips coverage for one line of test_workgroup_member: In
test_add_workgroup(), there is a chance that we will execute a `pop()`
of a set.  It all depends on if we get our expected item from the
two-item set.  Since the line's coverage is effectively random, we skip
coverage for it.