workgroup.tests: Two fields had errors
Two upstream error fields had incorrect identifiers in them (a wrong
SUNetID or a wrong name).  These are fields that are not examined by the
code, but they would be returned by the API, hence their inclusion.
Even though they're not looked at by the code, it's still good for the
test data to be accurate!