Workflow: Run update-version script on non-tag pushes
This changes the `main.yml` workflow, which runs on all pushes to the
repo.  It changes the `publish-test` job, which is the job responsible
for pushing releases to Test PyPi.  Specifically, the change *should*
cause the `update-version-for-test-push.py` script to run *only* when a
branch is being pushed, not when a tag is being pushed.

As per #34, tag pushes have been triggering development releases to Test
PyPi, at the same time that production releases are being pushed to
production PyPi.  That is not right: When a tag is being pushed, we want
production releases being published to both Test PyPi *and* production
PyPi.

Hopefully, this workflow change will cause the script to *not* be run on
a tag.

In order to test this, after this commit is pushed, I'll make another
commit to the changelog.  I'll push that, which will run the workflow
again.  Finally, I'll do a release, to ensure the new on-tag behavior
works.